refactor: raycast 폴더로 함수 옮김
