+++ config/packages.mk
@@ -11,15 +11,15 @@
 # **************************************************************************** #
 
 # @Packages
-PKGS = engine parser renderer
+PKGS = engine parser raycast renderer
 
-engineV = collision door init keyinput movement movement2 run
+engineV = collision door init keyinput movement movement_impl run
 parserV = check__cubfile check__cubid check__map check__map2\
 		check__map__index check__parser cubid fluidfill map map__raw\
 		parser rgb
+raycastV = floor raycast texture wall
 rendererV = camera dda__dist dda__step deinit draw image init load_tex\
-			minimap raycast raycast_floor raycast_texture raycast_tool\
-			raycast_wall
+			minimap util
 
 # Macro
 SRC_DIR = src/
