+++ include/engine.h
@@ -46,15 +46,15 @@ int				inputhandler__key_press(t_keycode key,
 /*
 ** < movement.c > */
 
+void			engine__move_player(t_engine *e);
+/*
+** < movement_impl.c > */
+
+t_mousemotion	is_mouse_motion(const t_inputhandler *this);
 void			engine__try_move_player(t_engine *e, const t_vec d);
 void			engine__try_move_player_straight(t_engine *e, t_sign dir);
 void			engine__try_strafe_player(t_engine *e, t_sign dir);
 void			engine__rotate_player(t_engine *e, double angle);
-t_mousemotion	is_mouse_motion(const t_inputhandler *this);
-/*
-** < movement2.c > */
-
-void			engine__move_player(t_engine *e);
 /*
 ** < run.c > */
