+++ src/raycast/texture.c
@@ -11,7 +11,7 @@
 /* ************************************************************************** */
 
 #include "engine.h"
-#include "renderer.h"
+#include "raycast.h"
 
 t_texdata	walldata__get_texdata(t_walldata *data, t_mapfmt tile_type)
 {
