+++ src/renderer/util.c
